Code Work/customer-support-classifier/requirements.txt
langgraph
langchain
langchain-openai
langchain-google-genai
gradio
pandas
python-dotenv
pillow
matplotlib
langchain-ollama
rich
ipython
